+++ tests/extensions/test_sat.py
@@ -1,13 +1,15 @@
 """Tests for pystac.extensions.sat."""
 
 import datetime
+from pystac.summaries import RangeSummary
 from typing import Any, Dict
 import unittest
 
 import pystac
+from pystac.utils import str_to_datetime, datetime_to_str
 from pystac import ExtensionTypeError
 from pystac.extensions import sat
-from pystac.extensions.sat import SatExtension
+from pystac.extensions.sat import OrbitState, SatExtension
 from tests.utils import TestCases
 
 
@@ -32,6 +34,21 @@ class SatTest(unittest.TestCase):
     def test_stac_extensions(self) -> None:
         self.assertTrue(SatExtension.has_extension(self.item))
 
+    def test_item_repr(self) -> None:
+        sat_item_ext = SatExtension.ext(self.item)
+        self.assertEqual(
+            f"<ItemSatExtension Item id={self.item.id}>", sat_item_ext.__repr__()
+        )
+
+    def test_asset_repr(self) -> None:
+        item = pystac.Item.from_file(self.sentinel_example_uri)
+        asset = item.assets["measurement_iw1_vh"]
+        sat_asset_ext = SatExtension.ext(asset)
+
+        self.assertEqual(
+            f"<AssetSatExtension Asset href={asset.href}>", sat_asset_ext.__repr__()
+        )
+
     def test_no_args_fails(self) -> None:
         SatExtension.ext(self.item).apply()
         with self.assertRaises(pystac.STACValidationError):
@@ -41,16 +58,45 @@ class SatTest(unittest.TestCase):
         orbit_state = sat.OrbitState.ASCENDING
         SatExtension.ext(self.item).apply(orbit_state)
         self.assertEqual(orbit_state, SatExtension.ext(self.item).orbit_state)
-        self.assertNotIn(sat.RELATIVE_ORBIT, self.item.properties)
-        self.assertFalse(SatExtension.ext(self.item).relative_orbit)
+        self.assertNotIn(sat.RELATIVE_ORBIT_PROP, self.item.properties)
+        self.assertIsNone(SatExtension.ext(self.item).relative_orbit)
         self.item.validate()
 
     def test_relative_orbit(self) -> None:
         relative_orbit = 1234
         SatExtension.ext(self.item).apply(None, relative_orbit)
         self.assertEqual(relative_orbit, SatExtension.ext(self.item).relative_orbit)
-        self.assertNotIn(sat.ORBIT_STATE, self.item.properties)
-        self.assertFalse(SatExtension.ext(self.item).orbit_state)
+        self.assertNotIn(sat.ORBIT_STATE_PROP, self.item.properties)
+        self.assertIsNone(SatExtension.ext(self.item).orbit_state)
+        self.item.validate()
+
+    def test_absolute_orbit(self) -> None:
+        absolute_orbit = 1234
+        SatExtension.ext(self.item).apply(absolute_orbit=absolute_orbit)
+        self.assertEqual(absolute_orbit, SatExtension.ext(self.item).absolute_orbit)
+        self.assertNotIn(sat.RELATIVE_ORBIT_PROP, self.item.properties)
+        self.assertIsNone(SatExtension.ext(self.item).relative_orbit)
+        self.item.validate()
+
+    def test_anx_datetime(self) -> None:
+        anx_datetime = str_to_datetime("2020-01-01T00:00:00Z")
+        SatExtension.ext(self.item).apply(anx_datetime=anx_datetime)
+        self.assertEqual(anx_datetime, SatExtension.ext(self.item).anx_datetime)
+        self.assertNotIn(sat.RELATIVE_ORBIT_PROP, self.item.properties)
+        self.assertIsNone(SatExtension.ext(self.item).relative_orbit)
+        self.item.validate()
+
+    def test_platform_international_designator(self) -> None:
+        platform_international_designator = "2018-080A"
+        SatExtension.ext(self.item).apply(
+            platform_international_designator=platform_international_designator
+        )
+        self.assertEqual(
+            platform_international_designator,
+            SatExtension.ext(self.item).platform_international_designator,
+        )
+        self.assertNotIn(sat.ORBIT_STATE_PROP, self.item.properties)
+        self.assertIsNone(SatExtension.ext(self.item).orbit_state)
         self.item.validate()
 
     def test_relative_orbit_no_negative(self) -> None:
@@ -104,8 +150,8 @@ class SatTest(unittest.TestCase):
         relative_orbit = 1002
         SatExtension.ext(self.item).apply(orbit_state, relative_orbit)
         d = self.item.to_dict()
-        self.assertEqual(orbit_state.value, d["properties"][sat.ORBIT_STATE])
-        self.assertEqual(relative_orbit, d["properties"][sat.RELATIVE_ORBIT])
+        self.assertEqual(orbit_state.value, d["properties"][sat.ORBIT_STATE_PROP])
+        self.assertEqual(relative_orbit, d["properties"][sat.RELATIVE_ORBIT_PROP])
 
         item = pystac.Item.from_dict(d)
         self.assertEqual(orbit_state, SatExtension.ext(item).orbit_state)
@@ -170,4 +216,116 @@ class SatTest(unittest.TestCase):
             r"^Satellite extension does not apply to type 'object'$",
             SatExtension.ext,
             object(),
+        )
+
+
+class SatSummariesTest(unittest.TestCase):
+    def setUp(self) -> None:
+        self.maxDiff = None
+
+    @staticmethod
+    def collection() -> pystac.Collection:
+        return pystac.Collection.from_file(
+            TestCases.get_path("data-files/collections/multi-extent.json")
+        )
+
+    def test_platform_international_designation(self) -> None:
+        collection = self.collection()
+        summaries_ext = SatExtension.summaries(collection)
+        platform_international_designator_list = ["2018-080A"]
+
+        summaries_ext.platform_international_designator = ["2018-080A"]
+
+        self.assertEqual(
+            summaries_ext.platform_international_designator,
+            platform_international_designator_list,
+        )
+
+        summaries_dict = collection.to_dict()["summaries"]
+
+        self.assertEqual(
+            summaries_dict["sat:platform_international_designator"],
+            platform_international_designator_list,
+        )
+
+    def test_orbit_state(self) -> None:
+        collection = self.collection()
+        summaries_ext = SatExtension.summaries(collection)
+        orbit_state_list = [OrbitState.ASCENDING]
+
+        summaries_ext.orbit_state = orbit_state_list
+
+        self.assertEqual(
+            summaries_ext.orbit_state,
+            orbit_state_list,
+        )
+
+        summaries_dict = collection.to_dict()["summaries"]
+
+        self.assertEqual(
+            summaries_dict["sat:orbit_state"],
+            orbit_state_list,
+        )
+
+    def test_absolute_orbit(self) -> None:
+        collection = self.collection()
+        summaries_ext = SatExtension.summaries(collection)
+        absolute_orbit_range = RangeSummary(2000, 3000)
+
+        summaries_ext.absolute_orbit = absolute_orbit_range
+
+        self.assertEqual(
+            summaries_ext.absolute_orbit,
+            absolute_orbit_range,
+        )
+
+        summaries_dict = collection.to_dict()["summaries"]
+
+        self.assertEqual(
+            summaries_dict["sat:absolute_orbit"],
+            absolute_orbit_range.to_dict(),
+        )
+
+    def test_relative_orbit(self) -> None:
+        collection = self.collection()
+        summaries_ext = SatExtension.summaries(collection)
+        relative_orbit_range = RangeSummary(50, 100)
+
+        summaries_ext.relative_orbit = relative_orbit_range
+
+        self.assertEqual(
+            summaries_ext.relative_orbit,
+            relative_orbit_range,
+        )
+
+        summaries_dict = collection.to_dict()["summaries"]
+
+        self.assertEqual(
+            summaries_dict["sat:relative_orbit"],
+            relative_orbit_range.to_dict(),
+        )
+
+    def test_anx_datetime(self) -> None:
+        collection = self.collection()
+        summaries_ext = SatExtension.summaries(collection)
+        anx_datetime_range = RangeSummary(
+            str_to_datetime("2020-01-01T00:00:00.000Z"),
+            str_to_datetime("2020-01-02T00:00:00.000Z"),
+        )
+
+        summaries_ext.anx_datetime = anx_datetime_range
+
+        self.assertEqual(
+            summaries_ext.anx_datetime,
+            anx_datetime_range,
+        )
+
+        summaries_dict = collection.to_dict()["summaries"]
+
+        self.assertDictEqual(
+            summaries_dict["sat:anx_datetime"],
+            {
+                "minimum": datetime_to_str(anx_datetime_range.minimum),
+                "maximum": datetime_to_str(anx_datetime_range.maximum),
+            },
         )